requirements.txt
appnope==0.1.0
asgiref==3.2.7
astroid==2.4.1
backcall==0.1.0
colorama==0.4.3
coverage==5.1
decorator==4.4.2
Django==3.0.7
django-cors-headers==3.4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
green==3.1.3
ipdb==0.13.2
ipython==7.14.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
psycopg2==2.8.5
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.2
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
traitlets==4.3.3
typed-ast==1.4.1
Unidecode==1.1.1
wcwidth==0.1.9
wrapt==1.12.1
